fix - Update investment amount handling to use BigDecimal for precision

+++ src/main/java/co/invest72/investment/console/input/reader/InstallmentInvestmentAmountReader.java
@@ -2,6 +2,7 @@ package co.invest72.investment.console.input.reader;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.math.BigDecimal;
 
 import co.invest72.investment.console.output.guide.GuidePrinter;
 import co.invest72.investment.domain.InvestmentAmount;
@@ -34,7 +35,7 @@ public class InstallmentInvestmentAmountReader implements InvestmentAmountReader
 			throw new IllegalArgumentException("투자 기간 단위는 '월' 또는 '년'이어야 합니다.");
 		} else if (periodType.equals("월")) {
 			int amount = Integer.parseInt(parts[1]);
-			return new MonthlyInstallmentInvestmentAmount(amount);
+			return new MonthlyInstallmentInvestmentAmount(BigDecimal.valueOf(amount));
 		} else {
 			int amount = Integer.parseInt(parts[1]);
 			return new YearlyInstallmentInvestmentAmount(amount);
